feat-Req-7-crie uma query que mostre uma relação de todos os álbuns produzidos por cada pessoa artista, com a quantidade de seguidores que ela possui, exibindo o nome pessoa artista, nome do album, quantidade pessoas seguindo

+++ desafio1.sql
@@ -5,7 +5,7 @@ CREATE DATABASE SpotifyClone;
 CREATE TABLE SpotifyClone.plans(
   plan_id INT NOT NULL AUTO_INCREMENT PRIMARY KEY,
   plan VARCHAR(30) NOT NULL,
-  price FLOAT
+  price DECIMAL(3,2)
 ) engine = InnoDB;
 
 INSERT INTO
